Fixed type mismatch between result_field and reference of integration sensor

+++ esphome/components/integration/integration_sensor.cpp
@@ -9,7 +9,7 @@ static const char *TAG = "integration";
 
 void IntegrationSensor::setup() {
   if (this->restore_) {
-    this->rtc_ = global_preferences.make_preference<float>(this->get_object_id_hash());
+    this->rtc_ = global_preferences.make_preference<double>(this->get_object_id_hash());
     this->rtc_.load(&this->result_);
   }
